bug(results): remove special characters before tokenizing
fulltext-search introduces the elastic search DSL which provides
the possibility to add magic characters. Those must be removed before
tokenizing for the result presentation.

This different DSL is annoying as it is the only one using it. The
other will just search the word with the magic characters and (usually)
return no results.

Closes: #46